+++ test/AdHoc.ZooKeeper.Tests/ZooKeeperTests.Create.cs
@@ -2,7 +2,7 @@ using AdHoc.ZooKeeper.Abstractions;
 
 namespace AdHoc.ZooKeeper.Tests;
 
-public abstract partial class ZooKeeperTests
+public partial class ZooKeeperTests
 {
 
     private static ZooKeeperPath _NewNode = "new-node";
