+++ test/AdHoc.ZooKeeper.Tests/ZooKeeperTests.cs
@@ -1,47 +1,105 @@
+using System.Collections.Frozen;
+using System.Collections.Immutable;
 using AdHoc.ZooKeeper.Abstractions;
+using DotNet.Testcontainers.Builders;
+using DotNet.Testcontainers.Containers;
+using DotNet.Testcontainers.Networks;
+using static AdHoc.ZooKeeper.Abstractions.ZooKeeperConnection;
 
 namespace AdHoc.ZooKeeper.Tests;
 
 [NotInParallel]
-public abstract partial class ZooKeeperTests
+public partial class ZooKeeperTests
 {
 
-    protected abstract IZooKeeper ZooKeeper { get; }
+    private const int _Instances = 3;
+    private static INetwork? _network;
+    private static List<IContainer> _containers = [];
 
-    protected abstract Task StartInstancesAsync(CancellationToken cancellationToken);
-    protected abstract Task StopInstancesAsync(CancellationToken cancellationToken);
-    protected abstract Task StopInstanceAsync(CancellationToken cancellationToken);
+    private static readonly SemaphoreSlim _lock = new(1, 1);
+    private static readonly Session _session = new(new Host("localhost"), FrozenSet<Authentication>.Empty, DefaultConnectionTimeout, DefaultSessionTimeout, false);
 
+    private ZooKeeperPath _root;
+    private ZooKeeper? _zoo;
+    protected IZooKeeper ZooKeeper => _zoo!;
 
-    [Test]
-    public Task PingAsync_WithConnection(CancellationToken cancellationToken) =>
-        ZooKeeper.PingAsync(cancellationToken);
-
-    [Test]
-    public async Task PingAsync_WithNoConnection(CancellationToken cancellationToken)
+    [Before(Class)]
+    public static async Task CreateZooAsync(CancellationToken cancellationToken)
     {
-        await StopInstancesAsync(cancellationToken);
-        await Assert.ThrowsAsync<ConnectionException>(() => ZooKeeper.PingAsync(cancellationToken));
+        _network = new NetworkBuilder()
+            .Build();
+        await _network.CreateAsync(cancellationToken);
+
+        for (var i = 0; i < _Instances; i++)
+            _containers.Add(CreateContainer(i + 1));
     }
 
-    [Test]
-    [DependsOn(nameof(PingAsync_WithConnection))]
-    public async Task PingAsync_WithLostConnection(CancellationToken cancellationToken)
-    {
-        await ZooKeeper.PingAsync(cancellationToken);
-        await StopInstancesAsync(cancellationToken);
-        await Assert.ThrowsAsync<ConnectionLostException>(() => ZooKeeper.PingAsync(cancellationToken));
-    }
+    private static IContainer CreateContainer(int i) =>
+        new ContainerBuilder()
+            .WithImage("zookeeper:latest")
+            .WithName($"{_network!.Name}-keeper{i}")
+            .WithEnvironment("ZOO_STANDALONE_ENABLED", "false")
+            .WithEnvironment("ZOO_MY_ID", i.ToString())
+            .WithEnvironment("ZOO_SERVERS", string.Join(' ', Enumerable.Range(1, _Instances).Select(j => $"server.{j}={_network!.Name}-keeper{j}:2888:3888;2181")))
+            .WithPortBinding(2181, true)
+            .WithWaitStrategy(Wait.ForUnixContainer().UntilPortIsAvailable(2181))
+            .WithNetwork(_network)
+            .Build();
 
-    [Test]
-    [DependsOn(nameof(PingAsync_WithLostConnection))]
-    public async Task PingAsync_WithReconnect(CancellationToken cancellationToken)
+    [Before(Test)]
+    public async Task PrepareContainerAsync(TestContext context, CancellationToken cancellationToken)
     {
-        await ZooKeeper.PingAsync(cancellationToken);
-        await StopInstancesAsync(cancellationToken);
-        await Assert.ThrowsAsync<ConnectionLostException>(() => ZooKeeper.PingAsync(cancellationToken));
+        _root = context.TestDetails.TestMethod.Name;
         await StartInstancesAsync(cancellationToken);
-        await ZooKeeper.PingAsync(cancellationToken);
+        await ZooKeeper.CreateAsync("/", cancellationToken);
     }
+
+    [After(Test)]
+    public async Task DisposeZooKeeperAsync(CancellationToken cancellationToken)
+    {
+        if (_zoo is not null)
+            await _zoo.DisposeAsync();
+    }
+
+    [After(Class)]
+    public static async Task DisposeZooAsync(CancellationToken cancellationToken)
+    {
+        foreach (var container in _containers)
+            await container.DisposeAsync();
+
+        if (_network is not null)
+            await _network.DisposeAsync();
+    }
+
+
+    private async Task StartInstancesAsync(CancellationToken cancellationToken)
+    {
+        await Task.WhenAll(_containers.Select(c => c.StartAsync(cancellationToken)));
+
+        ImmutableArray<Host> hosts = [.. _containers.Select(c => new Host("localhost", c.GetMappedPublicPort(2181)))];
+        _zoo = new ZooKeeper(_session, hosts, _root, _lock);
+        int i = 0;
+        while (i++ < 10)
+            try
+            {
+                if (!_session.IsConnected)
+                    await _session.ReconnectAsync(hosts[0], cancellationToken);
+                break;
+            }
+            catch
+            {
+                await Task.Delay(100);
+            }
+    }
+
+    private Task StopInstanceAsync(CancellationToken cancellationToken) =>
+        _containers.First(c => _session.Host.Port == c.GetMappedPublicPort(2181)).StopAsync(cancellationToken);
+
+    private async Task StopInstancesAsync(CancellationToken cancellationToken)
+    {
+        foreach (var container in _containers)
+            await container.StopAsync(cancellationToken);
+    }
+
 
 }
